requirements.txt
Flask==2.0.3
numpy==1.21.4
Pillow==8.4.0
requests==2.26.0
tensorflow==2.9.1
tensorflow-hub==0.12.0